+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21-1.4.3] - Unreleased
+### Added
+- Supplier Urn Upgrade. Loginar urns with this upgrade will replace consumed items (such as a tool or stack of blocks). For example, the urn could contain multiple stacks of planks, and it will replace each stack in your hand as you use them up.
+- Advancements
+### Fixed
+- Keybindings for Backpack and Swapper urns overriding other screens
+
 ## [1.21-1.4.2] - 2024-09-15
 ## Added
 - The potion pouch. It stores up to 9 potions and can be used to drink or throw them. It has a half-second cooldown after use to prevent accidental throws.